Remove unstable secondary tests
The ones about GUI and kvm and Xen.

Code did not changed in master nor in the testsuite and this tests are
failing when they weren't, which is the definition of unreliable :)

These were not failing in master, which is weird. Further inverstigation
might be needed to restore themin master, which is weird. Further
inverstigation might be needed to restore them.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>